setup main_common and main_prod
