+++ lib/constants.dart
@@ -1,3 +1,4 @@
+/// Environment variables and shared app constants.
 abstract class Constants {
   static const String supabaseUrl = String.fromEnvironment(
     'SUPABASE_URL',
