+++ lib/main_common.dart
@@ -1,25 +1,17 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_supabase_complete/app.dart';
 import 'package:flutter_supabase_complete/constants.dart';
 import 'package:supabase_flutter/supabase_flutter.dart';
 
-Future<void> main() async {
+/// Shared `runApp` configuration.
+///
+/// Used to initialize all required dependencies, packages, and constants.
+Future<void> mainCommon() async {
   WidgetsFlutterBinding.ensureInitialized();
   await Supabase.initialize(
     url: Constants.supabaseUrl,
     anonKey: Constants.supabaseAnnonKey,
   );
 
-  runApp(const MyApp());
-}
-
-class MyApp extends StatelessWidget {
-  const MyApp({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return const MaterialApp(
-      title: 'Flutter Supabase Complete',
-      home: Placeholder(),
-    );
-  }
+  runApp(const App());
 }
